requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
mysqlclient==2.2.6
PyMySQL==1.1.1
sqlparse==0.5.2
